Not overly complicated, should be part of init
